refactor: rename class CategoryProducts to ProductCategories and CategoryProductStats to ProductCategoryStats
